VIH-6177 remove judge available/unavailable events Judge joining does not start a hearing Judge transfer to hearing room now updates a conference to InSession
