+++ VideoAPI/VideoApi.AcceptanceTests/Features/Events.feature
@@ -30,8 +30,6 @@ Feature: Events
       | Pause                  |
       | Close                  |
       | Leave                  |
-      | JudgeAvailable         |
-      | JudgeUnavailable       |
       | MediaPermissionDenied  |
       | ParticipantJoining     |
       | Help                   |
