+++ VideoAPI/VideoApi.UnitTests/Events/TransferEventHandlerTests.cs
@@ -78,6 +78,46 @@ namespace VideoApi.UnitTests.Events
                     command.ConferenceId == conference.Id &&
                     command.ParticipantId == participantForEvent.Id &&
                     command.ParticipantState == It.IsAny<ParticipantState>())), Times.Never);
+            
+            CommandHandlerMock.Verify(
+                x => x.Handle(It.Is<UpdateConferenceStatusCommand>(command =>
+                    command.ConferenceId == conference.Id &&
+                    command.ConferenceState == ConferenceState.InSession)), Times.Never);
+        }
+        
+        [Test]
+        public async Task
+            Should_send_in_hearing_message_to_participants_and_live_message_to_service_bus_when_judge_transfers_to_hearing_room()
+        {
+            _eventHandler = new TransferEventHandler(QueryHandlerMock.Object, CommandHandlerMock.Object, RoomReservationService);
+
+            var conference = TestConference;
+            var participantForEvent = conference.GetParticipants().First(x => x.UserRole == UserRole.Judge);
+            
+            var callbackEvent = new CallbackEvent
+            {
+                EventType = EventType.Transfer,
+                EventId = Guid.NewGuid().ToString(),
+                ConferenceId = conference.Id,
+                ParticipantId = participantForEvent.Id,
+                TransferFrom = RoomType.WaitingRoom,
+                TransferTo = RoomType.HearingRoom,
+                TimeStampUtc = DateTime.UtcNow
+            };
+
+            await _eventHandler.HandleAsync(callbackEvent);
+
+            CommandHandlerMock.Verify(
+                x => x.Handle(It.Is<UpdateParticipantStatusAndRoomCommand>(command =>
+                    command.ConferenceId == conference.Id &&
+                    command.ParticipantId == participantForEvent.Id &&
+                    command.ParticipantState == ParticipantState.InHearing &&
+                    command.Room == RoomType.HearingRoom)), Times.Once);
+            
+            CommandHandlerMock.Verify(
+                x => x.Handle(It.Is<UpdateConferenceStatusCommand>(command =>
+                    command.ConferenceId == conference.Id &&
+                    command.ConferenceState == ConferenceState.InSession)), Times.Once);
         }
     }
 }
